Deprecated {sfsmisc}	R Documentation

Deprecated 'sfsmisc' Functions

Description

These functions are provided for compatibility with older versions of the sfsmisc package only, and may be defunct as soon as of the next release.

Usage

pmax.sa(scalar, arr)
pmin.sa(scalar, arr)

Arguments

scalar
numeric scalar.

arr
any numeric R object, typically array.

Details

pmax.sa(s, a) and pmin.sa(s, a) return (more-dimensional) arrays. These have been deprecated, because pmax and pmin do so too, if the array is used as first argument.

[Package sfsmisc version 1.1-3 Index]