cairoSwd {sfsmisc}	R Documentation

Cairo PDF Graphics Device useful for Sweave

Description

Provides a graphics device for Sweave, based on cairo_pdf. The advantage of cairoSwd() compared to pdf() is its support of Unicode characters.

Usage

cairoSwd(name, width, height, ...)

Arguments

name
file name prefix to which ‘.pdf’ will be appended.

width, height
in inches, see cairo_pdf.

...
further arguments, passed to cairo_pdf()

Note

Sweave devices need to have an argument list as above.

Usage in a Sweave chunk:

<<some-plot, fig=TRUE, grdevice=cairoSwd>>=

Author(s)

Alain Hauser

See Also

pdf, cairo_pdf, Sweave.

[Package sfsmisc version 1.1-3 Index]