col01scale {sfsmisc}	R Documentation

Matrix Scaling Utilities

Description

col01scale and colcenter (re)scale the columns of a matrix. These are simple one-line utilities, mainly with a didactical purpose.

Usage

colcenter (mat)
col01scale(mat, scale.func = function(x) diff(range(x)), location.func = mean)

Arguments

mat
numeric matrix, to rescaled.

scale.func, location.func
two functions mapping a numeric vector to a single number.

Value

a matrix with the same attributes as the input mat.

Author(s)

Martin Maechler

See Also

The standard R function scale().

Examples

## See the simple function definitions:

colcenter ## simply one line

col01scale# almost as simple

[Package sfsmisc version 1.1-12 Index]