ecdf.ksCI {sfsmisc}	R Documentation

Plot Empirical Distribution Function With 95% Confidence Curves

Description

Plots the empirical (cumulative) distribution function (ECDF) for univariate data, together with upper and lower simultaneous 95% confidence curves, computed via Kolmogorov-Smirnov' D, see KSd.

Usage

ecdf.ksCI(x, main = NULL, sub = NULL, xlab = deparse(substitute(x)),
          ci.col = "red", ...)

Arguments

x
x numerical vector of observations.

main,sub,xlab
arguments passed to title.

ci.col
color for confidence interval lines.

...
optional arguments passed to plot.stepfun.

Value

Nothing. Used for its side effect, to produce a plot.

Note

Presently, will only work if length(x) > 9.

Author(s)

Kjetil Halvorsen

References

Bickel and Doksum, see KSd.

See Also

ecdf and plot.stepfun in standard R.

Examples

ecdf.ksCI( rchisq(50,3) )

[Package sfsmisc version 1.1-12 Index]