empty.dimnames {sfsmisc}	R Documentation

Empty Dimnames of an Array

Description

Remove all dimension names from an array for compact printing.

Usage

empty.dimnames(a)

Arguments

a
an array, i.e., as special case a matrix.

Value

Returns a with its dimnames replaced by empty character strings.

Author(s)

Bill Venables / Martin Maechler, Sept 1993.

See Also

unname removes the dimnames.

Examples

empty.dimnames(diag(5)) # looks much nicer

(a <- matrix(-9:10, 4,5))
empty.dimnames(a) # nicer, right?

[Package sfsmisc version 1.1-12 Index]