f.robftest {sfsmisc}	R Documentation

Robust F-Test: Wald test for multiple coefficients of rlm() Object.

Description

Compute a robust F-Test, i.e., a Wald test for multiple coefficients of an rlm object.

Usage

f.robftest(object, var = -1)

Arguments

object
result of rlm().

var
variables. Either their names or their indices; the default, -1 means all but the intercept.

Details

This builds heavily on summary.rlm(), the summary method for rlm results.

Value

An object of class "htest", hence with the standard print methods for hypothesis tests. This is basically a list with components

statistic
the F statistic, according to ...

df
numerator and denominator degrees of freedom.

data.name
(extracted from input object.)

alternative
"two.sided", always.

p.value
the P-value, using an F-test on statistic and df[1:2].

Author(s)

Werner Stahel, July 2000; updates by Martin Maechler.

References

FIXME — Need some here !

See Also

rlm, summary.aov, etc.

Examples

if(require("MASS")) {
  ## same data as  example(rlm)
  data(stackloss)
  summary(rsl <- rlm(stack.loss ~ ., stackloss))
  f.robftest(rsl)
 } else  " forget it "

[Package sfsmisc version 1.1-12 Index]