helppdf {sfsmisc}	R Documentation

help() Type="pdf" and View It

Description

Utility to view PDF-rendered help pages; particularly useful in case they contain mathematical formulas or otherwise sophisticated formats.

Usage

helppdf(topic, viewer = getOption("pdfviewer"), quiet = !interactive(), ...)

Arguments

topic
the topic, passed to help().

viewer
a pdf viewer; the default is typically what you want interactively.

quiet
logical indicating that nothing should be printed to the console and the result should be returned as invisible().

...
further optional arguments passed to help().

Value

Returns the full path of the pdf file produced.

Author(s)

Martin Maechler

See Also

help, system.

Examples

## Both calls work :
if(interactive()) {
  helppdf(Normal)
  helppdf("NegBinomial")
} else { #  batch mode
  od <- setwd(tempdir())
  ff <- helppdf(Normal, viewer=NULL)
  stopifnot(file.exists(ff))
  setwd(od)# revert to previous dir.
}

[Package sfsmisc version 1.1-12 Index]