mpl {sfsmisc}	R Documentation

Simple Matrix Plots

Description

Do simple matrix plots, providing an easy interface to matplot by using a default x variable.

Usage

mpl(mat, ...)
p.m(mat, ...)

Arguments

mat
numeric matrix.

...
further arguments passed to matplot, e.g., type, xlab, etc.

Details

p.m(m) use the first column of m as x variable, whereas mpl(m) uses the integers 1, 2, ..., nrow(m) as coordinates and rownames(m) as axis labels if possible.

Note

These were really created for playing around with curves etc, and probably should be deprecated since in concrete examples, using matplot() directly is more appropriate.

Author(s)

Martin Maechler

See Also

matplot, plot.mts(*, plot.type = "single").

Examples

data(animals, package = "cluster")
mpl(animals, type = "l")

[Package sfsmisc version 1.1-12 Index]