n.plot {sfsmisc}	R Documentation

Name Plot: Names or Numbers instead of Points in Plot

Description

A utility function which basically calls plot(*, type="n") and text. To have names or numbers instead of points in a plot is useful for identifaction, e.g., in a residual plot, see also TA.plot.

Usage

n.plot(x, y = NULL, nam = NULL, abbr = n >= 20 || max(nchar(nam))>=8,
       xlab = NULL, ylab = NULL, log = "",
       cex = par("cex"), col = par("col"), ...)

Arguments

x,y
coordinates at which to plot. If y is missing, x is used for both, if it's a data.frame, list, 2-column matrix etc – via xy.coords; formula do not work.

nam
the labels to plot at each (x,y). Per default, these taken from the data x and y; case numbers 1:n are taken if no names are available.

abbr
logical indicating if the nam labels should be abbreviated – with a sensible default.

xlab,ylab
labels for the x- and y- axis, the latter being empty by default.

log
character specifying if log scaled axes should be used, see plot.default.

cex
plotting character expansion, see par.

col
color to use for text().

...
further arguments to be passed to the plot call.

Value

invisibly, a character vector with the labels used.

Author(s)

Martin Maechler, since 1992

See Also

plot.default, text.

Examples

 n.plot(1:20, cumsum(rnorm(20)))
 data(cars)
 with(cars, n.plot(speed, dist, cex = 0.8, col = "forest green"))

[Package sfsmisc version 1.1-3 Index]