p.arrows {sfsmisc}	R Documentation

Prettified Arrows Plots

Description

Draws arrows, like the arrows function, but with “nice” filled arrow heads.

Usage

p.arrows(x1, y1, x2, y2, size = 1, width, fill = 2, ...)

Arguments

x1, y1
coordinates of points from which to draw.

x2, y2
coordinates of points to which to draw.

size
symbol size as a fraction of a character height; default 1.

width
width of the arrow head; defaults to ....

fill
color for filling the arrow head.

...
further arguments passed to segments().

Author(s)

Andreas Ruckstuhl, 19 May 1994; (cosmetic by MM).

See Also

arrows.

Examples

example(arrows, echo = FALSE) #-> x, y, s
plot(x,y, main="p.arrows(.)")
p.arrows(x[s], y[s], x[s+1], y[s+1], col= 1:3, fill = "dark blue")

[Package sfsmisc version 1.1-7 Index]