p.hboxp {sfsmisc}	R Documentation

Add a Horizontal Boxplot to the Current Plot

Description

Add a horizontal boxplot to the current plot. This is mainly an auxiliary function for histBxp, since boxplot(*, horizontal = TRUE, add = TRUE) is usually much preferable to this.

Usage

p.hboxp(x, y.lo, y.hi, boxcol = 3,
        medcol = 2, medlwd = 5, whisklty = 2, staplelty = 1)

Arguments

x
univariate data set.

y.lo, y.hi
minimal and maximal user coordinates or y.lo = c(ylo,hyi).

boxcol, medcol
color of the box and the median line.

medlwd
line width of median line.

whisklty, staplelty
line types of the whisker and the staple, the latter being used for the outmost non-outliers.

Details

....

Author(s)

Martin Maechler building on code from Markus and Christian Keller.

See Also

boxplot(**, horizontal = TRUE, add= TRUE).

Examples

  ## ==>  See code in 'histBxp' (.)  and  example(histBxp) !
  ##

[Package sfsmisc version 1.1-3 Index]