p.profileTraces {sfsmisc}	R Documentation

Plot a profile.nls Object With Profile Traces

Description

Displays a series of plots of the profile t function and the likelihood profile traces for the parameters in a nonlinear regression model that has been fitted with nls and profiled with profile.nls.

Usage

p.profileTraces(x, cex = 1,
                subtitle = paste("t-Profiles and traces of ",
                       deparse(attr(x,"summary")$formula)))

Arguments

x
an object of class "profile.nls", typically resulting from profile(nls(.)), see profile.nls.

cex
character expansion, see par(cex =).

subtitle
a subtitle to set for the plot. The default now includes the nls() formula used.

Note

the stats-internal stats:::plot.profile.nls plot method just does “the diagonals”.

Author(s)

Andreas Ruckstuhl, R port by Isabelle Flückiger and Marcel Wolbers

See Also

profile, and nls (which has unexported profile and stats:::plot.profile.nls methods).

Examples

require(stats)
data(Puromycin)
Treat <- Puromycin[Puromycin$state == "treated", ]
fm <- nls(rate ~ T1*conc/(T2+conc), data=Treat,
          start = list(T1=207,T2=0.06))
(pr <- profile(fm)) # quite a few things..
op <- par(mfcol=1:2)
plot(pr) # -> 2 'standard' plots
par(op)
## ours:
p.profileTraces(pr)

[Package sfsmisc version 1.1-7 Index]