p.scales {sfsmisc}	R Documentation

Conversion between plotting scales: usr, cm, symbol

Description

Give scale conversion factors of three coordinate systems in use for traditional R graphics: use, cm, symbol.

Usage

p.scales(unit = relsysize * 2.54 * min(pin), relsysize = 0.05)

Arguments

unit
length of unit (or x and y units) of symbol coordinates in cm.

relsysize
same, as a proportion of the plotting area.

Value

A numeric 2x2 matrix, with rows named x and y, and columns, named "sy2usr" and "usr2cm" which give the scale conversion factors from ‘symbol’ (as given) to ‘usr’ coordinates and from these to ‘cm’, respectively.

Author(s)

Werner Stahel, 1990; simplification: M.Maechler, 1993, 2004

See Also

par("usr"), of also ("pin") on which this is based.

Examples

p.scales()

[Package sfsmisc version 1.1-3 Index]