potatoes {sfsmisc}	R Documentation

Fisher's Potato Crop Data

Description

Fisher's potato crop data set is of historical interest as an early example of a multi-factor block design.

Usage

data(potatoes)

Format

A data frame with 64 observations on the following 5 variables.

pos

a factor with levels 1:4.

treat

a factor with 16 levels A to H and J to Q, i.e., LETTERS[1:17][-9].

nitrogen

a factor specifying the amount of nitrogen sulfate (NH_4), with the four levels 0,1,2,4.

potash

a factor specifying the amount of potassium (K, ‘kalium’) sulfate, with the four levels 0,1,2,4.

yield

a numeric vector giving the yield of potatoes in ...

Source

Bennett, J. H. (1972) Collected Papers of R. A. Fischer vol.~II, 1925-31; The University of Adelaide.

References

T.Eden and R. A. Fisher (1929) Studies in Crop Variation. VI. Experiments on the Response of the Potato to Potash and Nitrogen. J. Agricultural Science 19, 201–213. Accessible from Bennett (1972), see above.

Examples

data(potatoes)
## See the experimental design:
with(potatoes, {
     cat("4 blocks of experiments;",
         "each does every (nitrogen,potash) combination (aka 'treat'ment) once.",
         '', sep="\n")
     print(ftable(table(nitrogen, potash, treat)))
     print(ftable(tt <- table(pos,potash,nitrogen)))
     tt[cbind(pos,potash,nitrogen)] <- as.character(treat)
     cat("The 4 blocks  pos = 1, 2, 3, 4:\n")
     ftable(tt)
     })
## First plot:
with(potatoes, interaction.plot(potash,nitrogen, response=yield))

## ANOVAs:
summary(aov(yield ~ nitrogen * potash + Error(pos), data = potatoes))
    # "==>" can use simply
summary(aov(yield ~ nitrogen + potash + pos, data = potatoes))
    # and
summary(aov(yield ~ nitrogen + potash, data = potatoes))

[Package sfsmisc version 1.1-3 Index]