quadrant {sfsmisc}	R Documentation

Give the Quadrant Number of Planar Points

Description

Determine the quadrant of planar points, i.e. in which of the four parts cut by the x- and y- axis the points lie. Zero values (i.e. points on the axes) are treated as if positive.

Usage

quadrant(x, y=NULL)

Arguments

x,y
numeric vectors of the same length, or x is an x-y structure and y=NULL, see xy.coords.

Value

numeric vector of same length as x (if that's a vector) with values in 1:4 indicating the quadrant number of the corresponding point.

Examples

xy <- as.matrix(expand.grid(x= -7:7, y= -7:7)); rownames(xy) <- NULL
(qu <- quadrant(xy))
plot(xy, col = qu+1, main = "quadrant() number", axes = FALSE)
abline(h=0, v=0, col="gray") # the x- and y- axis
text(xy, lab = qu, col = qu+1, adj = c(1.4,0))

[Package sfsmisc version 1.1-12 Index]