repChar {sfsmisc}	R Documentation

Make Simple String from Repeating a Character, e.g. Blank String

Description

Simple constructors of a constant character string from one character, notably a “blank” string of given string length.

M.M. is now ‘mentally deprecating’ bl.string in favor of using repChar() in all cases.

With R 3.3.0 (May 2016), the new function strrep() was introduced; it is faster typically, and more flexible, e.g. accepting a vector for the 2nd argument.
This (for now informally) deprecates all uses of repChar() and bl.string().

Usage

repChar(char, no)
bl.string(no)

Arguments

char
single character (or arbitrary string).

no
non-negative integer.

Value

One string, i.e., character(1)), for bl.string a blank string, fulfilling n == nchar(bl.string(n)).

Author(s)

Martin Maechler, early 1990's (for bl.string).

See Also

paste, character, nchar.

Examples

r <- sapply(0:8, function(n) ccat(repChar(" ",n), n))
cbind(r)

repChar("-", 4)
repChar("_", 6)
## it may make sense to a string of more than one character:
repChar("-=- ", 6)

## show the very simple function definitions:
repChar
bl.string

[Package sfsmisc version 1.1-12 Index]