rotn {sfsmisc}	R Documentation

Generalized Rot13 Character Translation (Rotation)

Description

Compute generalized ‘rot13’ character translations or “rotations”

In the distant past, considered as poor man's encryption, such rotations are way too poor nowadays and provided mainly for didactical reasons.

Usage

rotn(ch, n = 13)

Arguments

ch
a character vector; often a string (of length 1).

n
an integer in {1...26}; the default is particularly useful.

Details

Note that the default n = 13 makes rotn into a function that is its own inverse.

Written after having searched for it and found seqinr::rot13() which was generalized and rendered more transparently to my eyes.

Value

a character as ch, but with each character (which belongs to letters or LETTERS “rotated” by n (positions in the alphabet).

Author(s)

Martin Maechler

See Also

rot2, a completely different rotation (namely in the plane aka R^2).

Examples

rotn(c("ABC", "a","b","c"), 1)
rotn(c("ABC", "a","b","c"), 2)
rotn(c("ABC", "a","b","c"), 26) # rotation by 26 does not change much

(ch <- paste("Hello", c("World!", "you too")))
rotn(ch)
rotn( rotn(ch ) ) # rotn(*, 13) is its own inverse

[Package sfsmisc version 1.1-7 Index]