sessionInfoX {sfsmisc}	R Documentation

Extended Information About the Current R Session

Description

Collect (and print) information about the current R session and environment, using sessionInfo() and more mostly low-level and platform dependent information.

Usage

sessionInfoX(pkgs = NULL, list.libP = FALSE, extraR.env = TRUE)

## S3 method for class 'sessionInfoX'
print(x, locale = TRUE, RLIBS = TRUE, Renv = TRUE, ...)

Arguments

pkgs
an optional character vector of R packages, whose packageDescription()s are wanted.

list.libP
a logical indicating if for all .libPaths entries, the files should be listed via list.files.

extraR.env
logical indicating if all environment variables should be recorded which start with "R_" or "_R_".

x
typically the result of sessionInfoX().

locale
logical, passed to print.sessionInfo() indicating if the locale information should be printed.

RLIBS
logical indicating if the information about R_LIBS should be printed.

Renv
logical indicating if the information about R environment variables should be printed.

...
passed to print methods.

Value

an object of S3 class "sessionInfoX", a list with components (there may be more, experimental and not yet listed here):

sInfo
simply the value of sessionInfo().

sysInf
the value of Sys.info().

capabilities
the value of capabilities().

extSoft
for R 3.2.0 and newer, the value of extSoftVersion().

LAPACK
for R 3.0.3 and newer, the value of La_version().

pcre
for R 3.1.3 and newer, the value of pcre_config().

pkgDescr
If pkgs was non-empty, a named list of packageDescription()s for each entry in pkgs.

libPath
the value of .libPaths().

RLIBS
a character vector of entries from Sys.getenv("R_LIBS"), typically very similar to the libPaths component.

n.RLIBS
simply a normalizePath()ed version of RLIBS.

R.env
a named character vector with the “important” R environment variables R_ENVIRON, R_PROFILE, R_CHECK_ENVIRON.

xR.env
if extraR.env was true, a named character vector of “all R related” environment variables, as specified in extraR.env's description above.

Author(s)

Martin Maechler, December 2015

See Also

sessionInfo, .libPaths, R.version, Sys.getenv.

Examples

six0 <- sessionInfoX()
sixN <- sessionInfoX("nlme", list.libP = TRUE)
sixN # -> print() method for "sessionInfoX"
names(sixN)
str(sixN, max = 1)# outline of lower-level structure
str(sixN$pkgDescr) # list with one component "nlme"

[Package sfsmisc version 1.1-7 Index]