sfsmisc-defunct {sfsmisc}	R Documentation

Defunct Functions in Package sfsmisc

Description

The functions or variables listed here are no longer part of package sfsmisc as they are not needed (any more).

Usage

## Defunct in 2016-12 --> to be sfsmisc 1.1-1
list2mat(x, check = TRUE)

p.pllines(x,y,group,lty=c(1,3,2,4),...)

## deprecated from 2007 to 2013; defunct since 2014-01:
pl.ds() ##-- is replaced by  plotDS()

Details

list2mat(x) was usually the same as sapply(x, c) (where the latter does not construct column names where x has no names).

p.pllines is now defunct because basic R graphics (but not S-PLUS) provide its functionality directly: Use plot(x,y, lty = group, type = 'l', ...).

pl.ds has been renamed to plotDS() in 2007.

See Also

Defunct

[Package sfsmisc version 1.1-3 Index]