toLatex.numeric {sfsmisc}	R Documentation

LaTeX or Sweave friendly Formatting of Numbers

Description

Formats real numbers, possibly in scientific notation, with a given number of digits after the decimal point. Output can be used in LaTeX math mode, e.g., for printing numbers in a table, where each number has to be printed with the same number of digits after the decimal point, even if the last digits are zeros.

Usage

## S3 method for class 'numeric'
toLatex(object, digits = format.info(object)[2],
        scientific = format.info(object)[3] > 0, times = "\\cdot", ...)

Arguments

object
a numeric vector.

digits
number of digits after the decimal point (for the mantissa if scientific). The default behaves the same as R's format().

scientific
logical indicating if scientific notation a * 10^k should be used. The default behaves the same as R's format().

times
character string indicating the LaTeX symbol to be used for the ‘times’ sign.

...
unused; for compatibility with toLatex.

Value

a character vector of the same length as object, containing the formatted numbers.

Note

We use digits for round, i.e., round after the decimal point on purpose, rather than signif()icant digit rounding as used by print() or format().

Author(s)

Alain Hauser

See Also

pretty10exp which gives expressions similar to our scientific=TRUE. toLatex with other methods.

Examples

xx <- pi * 10^(-9:9)

format(xx)
formatC(xx)

toLatex(xx) #-> scientific = TRUE is chosen
toLatex(xx, scientific=FALSE)

sapply(xx, toLatex)
sapply(xx, toLatex, digits = 2)

[Package sfsmisc version 1.1-12 Index]