u.Datumvonheute {sfsmisc}	R Documentation

Datum und Uhrzeit (auf deutsch)

Description

Return current date and time as a string, possibly including day of the week in German.

Usage

u.Datumvonheute(W.tag=2, Zeit=FALSE)

C.Monatsname
C.Wochentag
C.Wochentagkurz
C.weekday

Arguments

W.tag
logical or integer specifying you want weekday (‘Wochentag’). 0 or FALSE gives no, 1 or TRUE gives a short and 2 the long version of the day of the week.

Zeit
logical or integer specifying if time ("Zeit") is desired. 0 or FALSE gives no, 1 or TRUE gives a hours only and 2 hours and minutes.

Value

A string with the current date/time, in the form specified by the arguments.

The C.* are character vector “constants”, the German ones actually used by u.Datumvonheute.

Author(s)

Caterina Savi, Martin Maechler

See Also

u.date for a similar English version, and p.datum which plots. For English month names, etc month.name.

Examples

u.Datumvonheute()
u.Datumvonheute(W.tag=1, Zeit=TRUE)
u.Datumvonheute(W.tag= FALSE, Zeit=2)

[Package sfsmisc version 1.1-12 Index]