u.boxplot.x {sfsmisc}	R Documentation

Utility Returning x-Coordinates of Boxplot

Description

Return the x-coordinates in an ‘n-way’ side-by-side boxplot. This is an auxiliary function and exists mainly for backcompatibility with S-plus.

Usage

u.boxplot.x(n, j = 1:n, fullrange = 100)

Arguments

n
number of boxplots.

j
indices of boxplots.

fullrange
x-coords as 'uniform' in [0,fullrange]; (f.=100, corresponds to Splus 3.x (x = 1,2)).

Value

a numeric vector of length n, with values inside (0,M) where M = fullrange.

Author(s)

Martin Maechler

See Also

boxplot.

Examples

 u.boxplot.x(7) # == 8.93 22.62 36.3 ... 91.07

[Package sfsmisc version 1.1-3 Index]