unif {sfsmisc}	R Documentation

Nice Uniform Points in Interval

Description

Give regularly spaced points on interval [-c,c] with mean 0 (exactly) and variance about 1 (very close for even n and larger round.dig). Note that c depends on n.

Usage

unif(n, round.dig = 1 + trunc(log10(n)))

Arguments

n
positive integer specifying the number of points desired.

round.dig
integer indicating to how many digits the result is rounded.

Value

numeric vector of length n, symmetric around 0, hence with exact mean 0, and variance approximately 1.

Note

It relies on the fact that Var(1,2,...,n) = n(n+1)/12.

Author(s)

Martin Maechler, ca 1990

See Also

runif for producing uniform random numbers.

Examples

(u <- unif(8))
var(u)

(u. <- unif(8, 12))# more digits in result, hence precision for Var :
var(u.)

[Package sfsmisc version 1.1-3 Index]